# libstore: Get rid of allocations in printString, allocate 2K bytes on the stack