@Ericson2314 2025-08-31
Bonus points if you can use std range or similar to iterate through chunks :)